requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.2
django-allauth==65.7.0
django-extensions==3.2.3
djangorestframework==3.16.0
idna==3.10
paypalrestsdk==1.13.3
pillow==11.1.0
pycparser==2.22
pyOpenSSL==25.0.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
